requirements.txt
streamlit>=1.25
pandas>=2.0
gspread>=5.7
google-auth>=2.25
